+++ backend/src/db/mod.rs
@@ -1,2 +1,3 @@
+pub mod database;
 pub mod models;
 pub mod schema;
